# Optimize assign op expression parsing